Extract "isFunction" assertion to own file
