+++ lib/referee.js
@@ -196,22 +196,7 @@ require("./assertions/defined")(referee);
 require("./assertions/is-null")(referee);
 require("./assertions/match")(referee);
 require("./assertions/is-object")(referee);
-
-referee.add("isFunction", {
-    assert: function (actual) {
-        return typeof actual === "function";
-    },
-    assertMessage: "${customMessage}${actual} (${actualType}) expected to be function",
-    refuteMessage: "${customMessage}${actual} expected not to be function",
-    expectation: "toBeFunction",
-    values: function (actual, message) {
-        return {
-            actual: String(actual).replace("\n", ""),
-            actualType: typeof actual,
-            customMessage: message
-        };
-    }
-});
+require("./assertions/is-function")(referee);
 
 referee.add("isTrue", {
     assert: function (actual) {
